@@ -1,8 +1,15 @@
 module github.com/ochinchina/supervisord/config
 
-go 1.16
+go 1.22
 
 require (
+	github.com/hashicorp/go-envparse v0.1.0
 	github.com/ochinchina/go-ini v1.0.1
-	github.com/sirupsen/logrus v1.8.1
+	github.com/ochinchina/supervisord/util v0.0.0-20230902082938-c2cae38b7454
+	github.com/sirupsen/logrus v1.9.3
+)
+
+require (
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
 )
