@@ -1,5 +1,10 @@
 module github.com/ochinchina/supervisord/events
 
-go 1.16
+go 1.22
 
-require github.com/sirupsen/logrus v1.8.1
+require github.com/sirupsen/logrus v1.9.3
+
+require (
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+)
