@@ -1,9 +1,10 @@
 module github.com/ochinchina/supervisord/faults
 
-go 1.16
+go 1.22
+
+require github.com/ochinchina/gorilla-xmlrpc v0.0.0-20171012055324-ecf2fe693a2c
 
 require (
-	github.com/gorilla/rpc v1.2.0 // indirect
-	github.com/ochinchina/gorilla-xmlrpc v0.0.0-20171012055324-ecf2fe693a2c
+	github.com/gorilla/rpc v1.2.1 // indirect
 	github.com/rogpeppe/go-charset v0.0.0-20190617161244-0dc95cdf6f31 // indirect
 )
