@@ -1,3 +1,17 @@
 module github.com/ochinchina/supervisord/logger
 
-go 1.16
+go 1.22
+
+require (
+	github.com/ochinchina/supervisord/events v0.0.0-20230902082938-c2cae38b7454
+	github.com/ochinchina/supervisord/faults v0.0.0-20230902082938-c2cae38b7454
+)
+
+require (
+	github.com/gorilla/rpc v1.2.1 // indirect
+	github.com/ochinchina/gorilla-xmlrpc v0.0.0-20171012055324-ecf2fe693a2c // indirect
+	github.com/rogpeppe/go-charset v0.0.0-20190617161244-0dc95cdf6f31 // indirect
+	github.com/sirupsen/logrus v1.9.3 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+)
