@@ -1,3 +1,3 @@
 module github.com/ochinchina/supervisord/types
 
-go 1.16
+go 1.22
