@@ -1,9 +1,13 @@
 module github.com/ochinchina/supervisord/xmlrpcclient
 
-go 1.16
+go 1.22
 
 require (
-	github.com/gorilla/rpc v1.2.0 // indirect
 	github.com/ochinchina/gorilla-xmlrpc v0.0.0-20171012055324-ecf2fe693a2c
+	github.com/ochinchina/supervisord/types v0.0.0-20230902082938-c2cae38b7454
+)
+
+require (
+	github.com/gorilla/rpc v1.2.1 // indirect
 	github.com/rogpeppe/go-charset v0.0.0-20190617161244-0dc95cdf6f31 // indirect
 )
