@@ -3,7 +3,6 @@ package config
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -117,7 +116,6 @@ func (c *Config) createEntry(name string, configDir string) *Entry {
 	return entry
 }
 
-//
 // Load the configuration and return loaded programs
 func (c *Config) Load() ([]string, error) {
 	myini := ini.NewIni()
@@ -141,7 +139,7 @@ func (c *Config) getIncludeFiles(cfg *ini.Ini) []string {
 			env := NewStringExpression("here", c.GetConfigFileDir())
 			files := strings.Fields(key)
 			for _, fRaw := range files {
-				dir := c.GetConfigFileDir()
+				var dir string
 				f, err := env.Eval(fRaw)
 				if err != nil {
 					continue
@@ -151,7 +149,7 @@ func (c *Config) getIncludeFiles(cfg *ini.Ini) []string {
 				} else {
 					dir = filepath.Join(c.GetConfigFileDir(), filepath.Dir(f))
 				}
-				fileInfos, err := ioutil.ReadDir(dir)
+				fileInfos, err := os.ReadDir(dir)
 				if err == nil {
 					goPattern := toRegexp(filepath.Base(f))
 					for _, fileInfo := range fileInfos {
@@ -405,7 +403,8 @@ func parseEnvFiles(s string) *map[string]string {
 }
 
 // GetEnv returns slice of strings with keys separated from values by single "=". An environment string example:
-//  environment = A="env 1",B="this is a test"
+//
+//	environment = A="env 1",B="this is a test"
 func (c *Entry) GetEnv(key string) []string {
 	value, ok := c.keyValues[key]
 	result := make([]string, 0)
@@ -426,7 +425,9 @@ func (c *Entry) GetEnv(key string) []string {
 }
 
 // GetEnvFromFiles returns slice of strings with keys separated from values by single "=". An envFile example:
-//  envFiles = global.env,prod.env
+//
+//	envFiles = global.env,prod.env
+//
 // cat global.env
 // varA=valueA
 func (c *Entry) GetEnvFromFiles(key string) []string {
@@ -512,7 +513,6 @@ func (c *Entry) GetStringArray(key string, sep string) []string {
 //	logSize=1GB
 //	logSize=1KB
 //	logSize=1024
-//
 func (c *Entry) GetBytes(key string, defValue int) int {
 	v, ok := c.keyValues[key]
 
